Fix WiFi AIDL removeNetwork() method
On L dessert, Google introduced a kind of "fake remove" for
WiFi networks in order to hide removed ones from the system
while keeping them in a configuration file:

    patch_ID: I12ea25142d0dcfc670d8970ed67624e8573d1ac2
    Author: vandwalle <vandwalle@google.com>
    Date: Sun Aug 31 23:26:39 2014 -0700

WiFi API offers two ways to delete a network: "removeNetwork"
and "forget" methods.

The "forget" method has been correctly updated to match the
above new patch but not the "removeNetwork" one.

If an app uses this public official documented "removeNetwork"
for a psk securised BSS, the removal will be partial and then,
it will never be possible to add again the same BSS.

This patch aligns the "removeNetwork" with the "forget" one to
comply with the recently added code.

Change-Id: I80d894d7d6888755cff7a6596c0ea4729d05c12d
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: Abdelmajid MLAYEH <abdelmajidx.mlayeh@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>